UI Add Table available actions on all components

+++ src/influxserver/influxservercfg.component.ts
@@ -8,7 +8,13 @@ import { ExportServiceCfg } from '../common/dataservice/export.service'
 
 import { GenericModal } from '../common/generic-modal';
 import { ExportFileModal } from '../common/dataservice/export-file-modal';
+import { Observable } from 'rxjs/Rx';
+
 import { ItemsPerPageOptions } from '../common/global-constants';
+import { TableActions } from '../common/table-actions';
+import { AvailableTableActions } from '../common/table-available-actions';
+
+declare var _:any;
 
 @Component({
   selector: 'influxservers',
@@ -46,8 +52,17 @@ export class InfluxServerCfgComponent {
     { title: 'User Agent', name: 'UserAgent' }
   ];
 
+  public tableAvailableActions : any;
+
+  editEnabled : boolean = false;
+  selectedArray : any = [];
+
+  public isRequesting : boolean;
+  public counterItems : number = null;
+  public counterErrors: any = [];
+
   public page: number = 1;
-  public itemsPerPage: number = 10;
+  public itemsPerPage: number = 20;
   public maxSize: number = 5;
   public numPages: number = 1;
   public length: number = 0;
@@ -65,6 +80,12 @@ export class InfluxServerCfgComponent {
     this.editmode = 'list';
     this.reloadData();
     this.builder = builder;
+  }
+
+  enableEdit() {
+    this.editEnabled = !this.editEnabled;
+    let obsArray = [];
+    this.tableAvailableActions = new AvailableTableActions('influxserver').availableOptions;
   }
 
   createStaticForm() {
@@ -89,6 +110,7 @@ export class InfluxServerCfgComponent {
     this.influxServerService.getInfluxServer(null)
       .subscribe(
       data => {
+        this.isRequesting = false;
         this.influxservers = data
         this.data = data;
         this.onChangeTable(this.config)
@@ -103,6 +125,25 @@ export class InfluxServerCfgComponent {
     this.myFilterValue = "";
     this.config.filtering = {filtering: { filterString: '' }};
     this.onChangeTable(this.config);
+  }
+
+  applyAction(test : any) : void {
+    switch(test.action) {
+       case "RemoveAllSelected": {
+          this.removeAllSelectedItems(this.selectedArray);
+          break;
+       }
+       case "ChangeProperty": {
+          this.updateAllSelectedItems(this.selectedArray,test.field,test.value)
+          break;
+       }
+       case "AppendProperty": {
+         this.updateAllSelectedItems(this.selectedArray,test.field,test.value,true);
+       }
+       default: {
+          break;
+       }
+    }
   }
 
   public changePage(page: any, data: Array<any> = this.data): Array<any> {
@@ -206,10 +247,6 @@ export class InfluxServerCfgComponent {
     this.length = sortedData.length;
   }
 
-  public onCellClick(data: any): any {
-    console.log(data);
-  }
-
   onFilter() {
     this.reloadData();
   }
@@ -221,6 +258,18 @@ export class InfluxServerCfgComponent {
 
   exportItem(item : any) : void {
     this.exportFileModal.initExportModal(item);
+  }
+
+  removeAllSelectedItems(myArray) {
+    let obsArray = [];
+    this.counterItems = 0;
+    this.isRequesting = true;
+    for (let i in myArray) {
+      console.log("Removing ",myArray[i].ID)
+      this.deleteInfluxServer(myArray[i].ID,true);
+      obsArray.push(this.deleteInfluxServer(myArray[i].ID,true));
+    }
+    this.genericForkJoin(obsArray);
   }
 
   removeItem(row) {
@@ -257,18 +306,27 @@ export class InfluxServerCfgComponent {
       );
  	}
 
-  deleteInfluxServer(id) {
+  deleteInfluxServer(id, recursive?) {
+    if (!recursive) {
     this.influxServerService.deleteInfluxServer(id)
       .subscribe(data => { },
       err => console.error(err),
       () => { this.viewModalDelete.hide(); this.editmode = "list"; this.reloadData() }
       );
+    } else {
+      return this.influxServerService.deleteInfluxServer(id)
+      .do(
+        (test) =>  { this.counterItems++},
+        (err) => { this.counterErrors.push({'ID': id, 'error' : err})}
+      );
+    }
   }
 
   cancelEdit() {
     this.editmode = "list";
     this.reloadData();
   }
+
   saveInfluxServer() {
     if (this.influxserverForm.valid) {
       this.influxServerService.addInfluxServer(this.influxserverForm.value)
@@ -281,19 +339,57 @@ export class InfluxServerCfgComponent {
     }
   }
 
-  updateInfluxServer() {
-    if (this.influxserverForm.valid) {
-      var r = true;
-      if (this.influxserverForm.value.ID != this.oldID) {
-        r = confirm("Changing Influx Server ID from " + this.oldID + " to " + this.influxserverForm.value.ID + ". Proceed?");
+  updateAllSelectedItems(mySelectedArray,field,value, append?) {
+    let obsArray = [];
+    this.counterItems = 0;
+    this.isRequesting = true;
+    if (!append)
+    for (let component of mySelectedArray) {
+      component[field] = value;
+      obsArray.push(this.updateInfluxServer(true,component));
+    } else {
+      let tmpArray = [];
+      if(!Array.isArray(value)) value = value.split(',');
+      console.log(value);
+      for (let component of mySelectedArray) {
+        console.log(value);
+        //check if there is some new object to append
+        let newEntries = _.differenceWith(value,component[field],_.isEqual);
+        tmpArray = newEntries.concat(component[field])
+        console.log(tmpArray);
+        component[field] = tmpArray;
+        obsArray.push(this.updateInfluxServer(true,component));
       }
-      if (r == true) {
-        this.influxServerService.editInfluxServer(this.influxserverForm.value, this.oldID)
-          .subscribe(data => { console.log(data) },
-          err => console.error(err),
-          () => { this.editmode = "list"; this.reloadData() }
-          );
+    }
+    this.genericForkJoin(obsArray);
+    //Make sync calls and wait the result
+    this.counterErrors = [];
+  }
+
+  updateInfluxServer(recursive?, component?) {
+    if(!recursive) {
+      if (this.influxserverForm.valid) {
+        var r = true;
+        if (this.influxserverForm.value.ID != this.oldID) {
+          r = confirm("Changing Influx Server ID from " + this.oldID + " to " + this.influxserverForm.value.ID + ". Proceed?");
+        }
+        if (r == true) {
+          this.influxServerService.editInfluxServer(this.influxserverForm.value, this.oldID)
+            .subscribe(data => { console.log(data) },
+            err => console.error(err),
+            () => { this.editmode = "list"; this.reloadData() }
+            );
+        }
       }
+    } else {
+      return this.influxServerService.editInfluxServer(component, component.ID)
+      .do(
+        (test) =>  { this.counterItems++ },
+        (err) => { this.counterErrors.push({'ID': component['ID'], 'error' : err['_body']})}
+      )
+      .catch((err) => {
+        return Observable.of({'ID': component.ID , 'error': err['_body']})
+      })
     }
   }
 
@@ -309,6 +405,17 @@ export class InfluxServerCfgComponent {
     () =>  { console.log("DONE")}
   );
 
+  }
+
+  genericForkJoin(obsArray: any) {
+    Observable.forkJoin(obsArray)
+              .subscribe(
+                data => {
+                  this.selectedArray = [];
+                  this.reloadData()
+                },
+                err => console.error(err),
+              );
   }
 
 }
