+++ src/measgroup/measgroupcfg.component.ts
@@ -6,10 +6,16 @@ import { InfluxMeasService } from '../influxmeas/influxmeascfg.service';
 import { ValidationService } from '../common/validation.service'
 import { FormArray, FormGroup, FormControl} from '@angular/forms';
 import { ExportServiceCfg } from '../common/dataservice/export.service'
+import { Observable } from 'rxjs/Rx';
 
 import { GenericModal } from '../common/generic-modal';
 import { ExportFileModal } from '../common/dataservice/export-file-modal';
 import { ItemsPerPageOptions } from '../common/global-constants';
+
+import { TableActions } from '../common/table-actions';
+import { AvailableTableActions } from '../common/table-available-actions';
+
+declare var _:any;
 
 @Component({
   selector: 'measgroups',
@@ -22,6 +28,12 @@ export class MeasGroupCfgComponent {
   @ViewChild('viewModal') public viewModal: GenericModal;
   @ViewChild('viewModalDelete') public viewModalDelete: GenericModal;
   @ViewChild('exportFileModal') public exportFileModal : ExportFileModal;
+
+  editEnabled : boolean = false;
+  selectedArray : any = [];
+  public isRequesting : boolean;
+  public counterItems : number = null;
+  public counterErrors: any = [];
 
   itemsPerPageOptions : any = ItemsPerPageOptions;
   editmode: string; //list , create, modify
@@ -41,8 +53,9 @@ export class MeasGroupCfgComponent {
     { title: 'Measurements', name: 'Measurements' },
   ];
 
+  public tableAvailableActions : any;
   public page: number = 1;
-  public itemsPerPage: number = 10;
+  public itemsPerPage: number = 20;
   public maxSize: number = 5;
   public numPages: number = 1;
   public length: number = 0;
@@ -63,6 +76,14 @@ export class MeasGroupCfgComponent {
     this.builder = builder;
   }
 
+  enableEdit() {
+    this.editEnabled = !this.editEnabled;
+    console.log(this.editEnabled);
+    let obsArray = [];
+    this.tableAvailableActions = new AvailableTableActions('measgroup').availableOptions;
+    console.log(this.tableAvailableActions);
+  }
+
   createStaticForm() {
     this.measgroupForm = this.builder.group({
       ID: [this.measgroupForm ? this.measgroupForm.value.ID : '', Validators.required],
@@ -72,10 +93,13 @@ export class MeasGroupCfgComponent {
   }
 
   reloadData() {
+    this.selectedArray = [];
+    this.isRequesting = true;
     // now it's a simple subscription to the observable
     this.measGroupService.getMeasGroup(null)
       .subscribe(
       data => {
+        this.isRequesting = false;
         this.measgroups = data;
         this.data = data;
         this.onChangeTable(this.config);
@@ -90,6 +114,18 @@ export class MeasGroupCfgComponent {
     this.myFilterValue = "";
     this.config.filtering = {filtering: { filterString: '' }};
     this.onChangeTable(this.config);
+  }
+
+  applyAction(test : any) : void {
+    switch(test.action) {
+       case "RemoveAllSelected": {
+          this.removeAllSelectedItems(this.selectedArray);
+          break;
+       }
+       default: {
+          break;
+       }
+    }
   }
 
   public changePage(page: any, data: Array<any> = this.data): Array<any> {
@@ -193,10 +229,6 @@ export class MeasGroupCfgComponent {
     this.length = sortedData.length;
   }
 
-  public onCellClick(data: any): any {
-    console.log(data);
-  }
-
   onFilter() {
     this.reloadData();
   }
@@ -208,6 +240,18 @@ export class MeasGroupCfgComponent {
 
   exportItem(item : any) : void {
     this.exportFileModal.initExportModal(item);
+  }
+
+  removeAllSelectedItems(myArray) {
+    let obsArray = [];
+    this.counterItems = 0;
+    this.isRequesting = true;
+    for (let i in myArray) {
+      console.log("Removing ",myArray[i].ID)
+      this.deleteMeasGroup(myArray[i].ID,true);
+      obsArray.push(this.deleteMeasGroup(myArray[i].ID,true));
+    }
+    this.genericForkJoin(obsArray);
   }
 
   removeItem(row) {
@@ -248,12 +292,20 @@ export class MeasGroupCfgComponent {
       );
   }
 
-  deleteMeasGroup(id) {
+  deleteMeasGroup(id, recursive?) {
+    if (!recursive) {
     this.measGroupService.deleteMeasGroup(id)
       .subscribe(data => { },
       err => console.error(err),
       () => { this.viewModalDelete.hide(); this.editmode = "list"; this.reloadData() }
       );
+    } else {
+      return this.measGroupService.deleteMeasGroup(id)
+      .do(
+        (test) =>  { this.counterItems++},
+        (err) => { this.counterErrors.push({'ID': id, 'error' : err})}
+      );
+    }
   }
   cancelEdit() {
     this.reloadData();
@@ -269,19 +321,57 @@ export class MeasGroupCfgComponent {
     }
   }
 
-  updateMeasGroup() {
-    if (this.measgroupForm.valid) {
-      var r = true;
-      if (this.measgroupForm.value.ID != this.oldID) {
-        r = confirm("Changing Measurement Group ID from " + this.oldID + " to " + this.measgroupForm.value.ID + ". Proceed?");
+  updateAllSelectedItems(mySelectedArray,field,value, append?) {
+    let obsArray = [];
+    this.counterItems = 0;
+    this.isRequesting = true;
+    if (!append)
+    for (let component of mySelectedArray) {
+      component[field] = value;
+      obsArray.push(this.updateMeasGroup(true,component));
+    } else {
+      let tmpArray = [];
+      if(!Array.isArray(value)) value = value.split(',');
+      console.log(value);
+      for (let component of mySelectedArray) {
+        console.log(value);
+        //check if there is some new object to append
+        let newEntries = _.differenceWith(value,component[field],_.isEqual);
+        tmpArray = newEntries.concat(component[field])
+        console.log(tmpArray);
+        component[field] = tmpArray;
+        obsArray.push(this.updateMeasGroup(true,component));
       }
-      if (r == true) {
-        this.measGroupService.editMeasGroup(this.measgroupForm.value, this.oldID)
-          .subscribe(data => { console.log(data) },
-          err => console.error(err),
-          () => { this.editmode = "list"; this.reloadData() }
-          );
+    }
+    this.genericForkJoin(obsArray);
+    //Make sync calls and wait the result
+    this.counterErrors = [];
+  }
+
+  updateMeasGroup(recursive?, component?) {
+    if (!recursive) {
+      if (this.measgroupForm.valid) {
+        var r = true;
+        if (this.measgroupForm.value.ID != this.oldID) {
+          r = confirm("Changing Measurement Group ID from " + this.oldID + " to " + this.measgroupForm.value.ID + ". Proceed?");
+        }
+        if (r == true) {
+          this.measGroupService.editMeasGroup(this.measgroupForm.value, this.oldID)
+            .subscribe(data => { console.log(data) },
+            err => console.error(err),
+            () => { this.editmode = "list"; this.reloadData() }
+            );
+        }
       }
+    } else {
+      return this.measGroupService.editMeasGroup(component, component.ID)
+      .do(
+        (test) =>  { this.counterItems++ },
+        (err) => { this.counterErrors.push({'ID': component['ID'], 'error' : err['_body']})}
+      )
+      .catch((err) => {
+        return Observable.of({'ID': component.ID , 'error': err['_body']})
+      })
     }
   }
 
@@ -299,6 +389,17 @@ export class MeasGroupCfgComponent {
       err => console.error(err),
       () => { console.log('DONE') }
       );
+  }
+
+  genericForkJoin(obsArray: any) {
+    Observable.forkJoin(obsArray)
+              .subscribe(
+                data => {
+                  this.selectedArray = [];
+                  this.reloadData()
+                },
+                err => console.error(err),
+              );
   }
 
 }
